Moved error.js to handlers and attached error handler to router
