+++ index.js
@@ -33,7 +33,7 @@ var deleteHandler = require('./lib/handlers/delete.js');
 var patchHandler = require('./lib/handlers/patch.js');
 
 // Error page handler
-var errorHandler = require('./lib/error.js');
+var errorHandler = require('./lib/handlers/error.js');
 
 function ldnode (argv) {
     var ldp = new LDP(argv);
@@ -72,11 +72,6 @@ function createServer(argv) {
     var ldpApp = ldnode(argv);
     var ldp = ldpApp.locals.ldp;
     app.use(ldp.mount, ldpApp);
-
-    //Error handling
-    app.use(errorHandler.handler);
-
-    app.locals.ldp = ldp;
 
     if (ldp && (ldp.webid || ldp.key || ldp.cert) ) {
         debug.settings("SSL Private Key path: " + ldp.key);
@@ -196,6 +191,9 @@ function routes () {
     router.delete('/*', deleteHandler.handler);
     router.post('/*', postHandler.handler);
     router.patch('/*', patchHandler.handler);
+
+    //Error handling
+    router.use(errorHandler.handler);
     return router;
 }
