+++ lib/handlers/error.js
